update the product using put method

+++ src/Components/Pages/Cart/Cart.jsx
@@ -4,37 +4,55 @@ import { useState } from "react";
 import Swal from "sweetalert2";
 
 const Cart = () => {
+  const loadedProducts = useLoaderData();
+  const [products, setProducts] = useState(loadedProducts);
 
-    const loadedProducts = useLoaderData();
-    const [products, setProducts] = useState(loadedProducts);
-
-    const handelDelete = id =>{
-        console.log(id);
+  const handelDelete = (id) => {
+    console.log(id);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
         fetch(`http://localhost:5000/cart/${id}`, {
-            method: 'DELETE'
+          method: "DELETE",
         })
-        .then(res => res.json())
-        .then(data => {
-            if(data.deletedCount > 0){
-                const remainingProducts = products.filter(product => product._id !== id);
-                setProducts(remainingProducts);
-                Swal.fire({
-                    icon: 'success',
-                    title: 'Product deleted Successfully',
-                    showConfirmButton: false,
-                    timer: 1500
-                  })
-            }
-        })
-    }
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              const remainingProducts = products.filter(
+                (product) => product._id !== id
+              );
+              setProducts(remainingProducts);
 
-    return (
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-5">
-            {
-                products.map(product => <DisplayCartProducts key={product._id} product={product} handelDelete={handelDelete} />)
+              Swal.fire({
+                icon: "success",
+                title: "Product has been deleted.",
+                showConfirmButton: false,
+                timer: 1000,
+              });
             }
-        </div>
-    );
+          });
+      }
+    });
+  };
+
+  return (
+    <div className="grid grid-cols-1 md:grid-cols-2 gap-5">
+      {products.map((product) => (
+        <DisplayCartProducts
+          key={product._id}
+          product={product}
+          handelDelete={handelDelete}
+        />
+      ))}
+    </div>
+  );
 };
 
 export default Cart;
